+++ packages/kitchensink/service-fastify/feature-register-route-get.js
@@ -13,11 +13,19 @@ module.exports = ({ registerAction }) => {
    * Using the `registerRoute` API with positional arguments
    */
   registerAction({
-    target: '$FASTIFY_GET',
+    target: '$FASTIFY_ROUTE',
     name: 'registerRouteGet',
     handler: ({ registerRoute }) => {
-      registerRoute('/info1', async () => 'Info1');
-      registerRoute('/info2', async () => 'Info2');
+      registerRoute({
+        method: 'GET',
+        url: '/info1',
+        handler: async () => 'Info1',
+      });
+      registerRoute({
+        method: 'GET',
+        url: '/info2',
+        handler: async () => 'Info2',
+      });
     },
   });
   /**
@@ -26,55 +34,63 @@ module.exports = ({ registerAction }) => {
    */
 
   registerAction({
-    target: '$FASTIFY_GET',
+    target: '$FASTIFY_ROUTE',
     name: 'registerRouteGet_returningSingle',
     handler: () => ({
+      method: 'GET',
       url: '/info3',
       handler: async () => 'Info3',
     }),
   });
+
   /**
    * Using the returning API
    * (multiple values)
    */
 
   registerAction({
-    target: '$FASTIFY_GET',
+    target: '$FASTIFY_ROUTE',
     name: 'registerRouteGet_returningMulti',
     handler: () => [
       {
+        method: 'GET',
         url: '/info4',
         handler: async () => 'Info4',
       },
       {
+        method: 'GET',
         url: '/info5',
         handler: async () => 'Info5',
       },
     ],
   });
+
   /**
    * Using the declarative API
    * (single value)
    */
 
   registerAction({
-    target: '$FASTIFY_GET',
+    target: '$FASTIFY_ROUTE',
     name: 'registerRouteGet_declarativeSingle',
     handler: {
+      method: 'GET',
       url: '/info6',
       handler: async () => 'Info6',
     },
   });
+
   /**
    * Using the declarative API
    * (multiple values)
    */
 
   registerAction({
-    target: '$FASTIFY_GET',
+    target: '$FASTIFY_ROUTE',
     name: 'registerRouteGet_declarativeMulti',
     handler: [
       {
+        method: 'GET',
         url: '/info7',
         handler: async () => 'Info7',
       },